@@ -12,7 +12,7 @@ classifiers =
     Programming Language :: Python :: 3.10
 
 [options]
-python_requires = >=3.8
+python_requires = >=3.9
 packages = find:
 package_dir =
     =src
