@@ -1,5 +1,5 @@
 from enum import Enum
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, Optional
 
 import numpy as np
 from pydantic import BaseModel, validator
@@ -66,7 +66,7 @@ class IspybParams(BaseModel):
     comment: str
     resolution: float
 
-    sample_id: Optional[int] = None
+    sample_id: Optional[str] = None
     sample_barcode: Optional[str] = None
 
     # Optional from GDA as populated by Ophyd
@@ -75,8 +75,8 @@ class IspybParams(BaseModel):
     synchrotron_mode: Optional[str] = None
     slit_gap_size_x: Optional[float] = None
     slit_gap_size_y: Optional[float] = None
-    xtal_snapshots_omega_start: Optional[List[str]] = None
-    xtal_snapshots_omega_end: Optional[List[str]] = None
+    xtal_snapshots_omega_start: Optional[list[str]] = None
+    xtal_snapshots_omega_end: Optional[list[str]] = None
 
     @validator("transmission_fraction")
     def _transmission_not_percentage(cls, transmission_fraction: float):
