@@ -64,6 +64,7 @@ def pin_centre_then_xray_centre_plan(
         parameters.experiment_params.tip_offset_microns,
         oav_config_file,
     )
+
     grid_detect_params = create_parameters_for_grid_detection(parameters)
 
     oav_params = OAVParameters("xrayCentring", oav_config_file)
@@ -78,6 +79,7 @@ def pin_centre_then_xray_centre_plan(
 def pin_tip_centre_then_xray_centre(
     composite: GridDetectThenXRayCentreComposite,
     parameters: PinCentreThenXrayCentreInternalParameters,
+    oav_config_files: dict[str, str] = OAV_CONFIG_FILE_DEFAULTS,
 ) -> MsgGenerator:
     """Starts preparing for collection then performs the pin tip centre and xray centre"""
 
@@ -87,5 +89,7 @@ def pin_tip_centre_then_xray_centre(
 
     return start_preparing_data_collection_then_do_plan(
         eiger,
-        pin_centre_then_xray_centre_plan(composite, parameters),
+        composite.detector_motion,
+        parameters.experiment_params.detector_distance,
+        pin_centre_then_xray_centre_plan(composite, parameters, oav_config_files),
     )
