@@ -3,8 +3,12 @@ from __future__ import annotations
 from hyperion.external_interaction.callbacks.ispyb_callback_base import (
     BaseISPyBCallback,
 )
-from hyperion.external_interaction.ispyb.store_in_ispyb import StoreRotationInIspyb
+from hyperion.external_interaction.ispyb.store_in_ispyb import (
+    IspybIds,
+    StoreRotationInIspyb,
+)
 from hyperion.log import ISPYB_LOGGER, set_dcgid_tag
+from hyperion.parameters.constants import ROTATION_OUTER_PLAN, ROTATION_PLAN_MAIN
 from hyperion.parameters.plan_specific.rotation_scan_internal_params import (
     RotationInternalParameters,
 )
@@ -27,27 +31,29 @@ class RotationISPyBCallback(BaseISPyBCallback):
     Usually used as part of a RotationCallbackCollection.
     """
 
-    def __init__(self, parameters: RotationInternalParameters):
-        self.params: RotationInternalParameters
-        super().__init__(parameters)
-        self.ispyb: StoreRotationInIspyb = StoreRotationInIspyb(
-            self.ispyb_config, self.params
-        )
-        self.ispyb_ids: tuple[int, int] | tuple[None, None] = (None, None)
-
     def append_to_comment(self, comment: str):
-        assert self.ispyb_ids[0] is not None
-        self._append_to_comment(self.ispyb_ids[0], comment)
+        assert isinstance(self.ispyb_ids.data_collection_ids, int)
+        self._append_to_comment(self.ispyb_ids.data_collection_ids, comment)
 
-    def start(self, doc: dict):
+    def activity_gated_start(self, doc: dict):
+        if doc.get("subplan_name") == ROTATION_OUTER_PLAN:
+            ISPYB_LOGGER.info(
+                "ISPyB callback recieved start document with experiment parameters."
+            )
+            json_params = doc.get("hyperion_internal_parameters")
+            self.params = RotationInternalParameters.from_json(json_params)
+            self.ispyb: StoreRotationInIspyb = StoreRotationInIspyb(
+                self.ispyb_config, self.params
+            )
+        self.ispyb_ids: IspybIds = IspybIds()
         ISPYB_LOGGER.info("ISPYB handler received start document.")
-        if doc.get("subplan_name") == "rotation_scan_main":
+        if doc.get("subplan_name") == ROTATION_PLAN_MAIN:
             self.uid_to_finalize_on = doc.get("uid")
 
-    def event(self, doc: dict):
-        super().event(doc)
-        set_dcgid_tag(self.ispyb_ids[1])
+    def activity_gated_event(self, doc: dict):
+        super().activity_gated_event(doc)
+        set_dcgid_tag(self.ispyb_ids.data_collection_group_id)
 
-    def stop(self, doc: dict):
+    def activity_gated_stop(self, doc: dict):
         if doc.get("run_start") == self.uid_to_finalize_on:
-            super().stop(doc)
+            super().activity_gated_stop(doc)
