@@ -5,11 +5,13 @@ from hyperion.external_interaction.callbacks.ispyb_callback_base import (
 )
 from hyperion.external_interaction.exceptions import ISPyBDepositionNotMade
 from hyperion.external_interaction.ispyb.store_in_ispyb import (
+    IspybIds,
     Store2DGridscanInIspyb,
     Store3DGridscanInIspyb,
     StoreGridscanInIspyb,
 )
 from hyperion.log import ISPYB_LOGGER, set_dcgid_tag
+from hyperion.parameters.constants import GRIDSCAN_OUTER_PLAN
 from hyperion.parameters.plan_specific.gridscan_internal_params import (
     GridscanInternalParameters,
 )
@@ -32,30 +34,42 @@ class GridscanISPyBCallback(BaseISPyBCallback):
     Usually used as part of an FGSCallbackCollection.
     """
 
-    def __init__(self, parameters: GridscanInternalParameters):
-        super().__init__(parameters)
+    def __init__(self) -> None:
+        super().__init__()
         self.params: GridscanInternalParameters
-        self.ispyb: StoreGridscanInIspyb = (
-            Store3DGridscanInIspyb(self.ispyb_config, self.params)
-            if self.params.experiment_params.is_3d_grid_scan
-            else Store2DGridscanInIspyb(self.ispyb_config, self.params)
-        )
-        self.ispyb_ids: tuple = (None, None, None)
+        self.ispyb: StoreGridscanInIspyb
+        self.ispyb_ids: IspybIds = IspybIds()
 
-    def append_to_comment(self, comment: str):
-        for id in self.ispyb_ids[0]:
-            self._append_to_comment(id, comment)
+    def activity_gated_start(self, doc: dict):
+        if doc.get("subplan_name") == GRIDSCAN_OUTER_PLAN:
+            self.uid_to_finalize_on = doc.get("uid")
+            ISPYB_LOGGER.info(
+                "ISPyB callback recieved start document with experiment parameters and"
+                f"uid: {self.uid_to_finalize_on}"
+            )
+            json_params = doc.get("hyperion_internal_parameters")
+            self.params = GridscanInternalParameters.from_json(json_params)
+            self.ispyb = (
+                Store3DGridscanInIspyb(self.ispyb_config, self.params)
+                if self.params.experiment_params.is_3d_grid_scan
+                else Store2DGridscanInIspyb(self.ispyb_config, self.params)
+            )
 
-    def event(self, doc: dict):
-        super().event(doc)
-        set_dcgid_tag(self.ispyb_ids[2])
+    def activity_gated_event(self, doc: dict):
+        super().activity_gated_event(doc)
+        set_dcgid_tag(self.ispyb_ids.data_collection_group_id)
 
-    def stop(self, doc: dict):
+    def activity_gated_stop(self, doc: dict):
         if doc.get("run_start") == self.uid_to_finalize_on:
             ISPYB_LOGGER.info(
-                "ISPyB callback received stop document corresponding to start document"
-                "uid."
+                "ISPyB callback received stop document corresponding to start document "
+                f"with uid: {self.uid_to_finalize_on}."
             )
-            if self.ispyb_ids == (None, None, None):
+            if self.ispyb_ids == IspybIds():
                 raise ISPyBDepositionNotMade("ispyb was not initialised at run start")
-            super().stop(doc)
+            super().activity_gated_stop(doc)
+
+    def append_to_comment(self, comment: str):
+        assert isinstance(self.ispyb_ids.data_collection_ids, tuple)
+        for id in self.ispyb_ids.data_collection_ids:
+            self._append_to_comment(id, comment)
